Consistent checking of __GNUC__ define.
The kj/common.h header sometimes checks for if __GNUC__ is defined, and sometimes if it's true, this is changing it to be consistent in checking if it's true.

Specifically, MSVC classifies checking the value of an undefined define as a warning, this header can be included in other projects that have warnings as errors turned on.

My approach to solving this was to manually define these defines as 0, such as ___GNUC__, __CLANG__, __NO_INLINE__, KJ_USE_FUTEX, etc. However doing such for __GNUC__ caused additional problems due to this inconcinstency when it is defined, but set to 0.

This could alternativelly be fixed by consistently checking for all the macro definitions being undefined.